Increase net dosh rewards as if faked players were real
Prior to this commit, zed dosh scaled down with fake players so that
the total amount of kill-reward dosh a solo player earned would be
constant regardless of the faked player count.  This meant that you
would get 1 dosh per clot on five faked solo.  Getting through a 190
zed wave 1 on five faked solo is hard enough; trying to clear wave 2
with ~760 dosh is harder, and impossible for some classes/perks.

Now, you get the same amount of dosh per-kill that you would if the
faked players were all real.  So, on a medium solo game with 5 faked,
each clot gives 8 dosh, each gorefast gives 14, etc up to each fp
giving 240 dosh.  On a long game, it would be 7 per clot up to 192
per fp.  These are the same numbers you would get in online play with
five other humans, except that it all goes to you in solo mode.