Refining nondeterministic spawn debug cmds
Added a pair of console commands to help inspect the effects
of a customized spawn cycle.

CDSpawnDetails prints each wave's spawn cycle on a separate line.

CDSpawnSummaries, which takes an optional number of players or
tries to guess the correct number of players using FakePlayers if
no argument is given, prints the number of FP, SC, Husk, Siren,
Bloat, and trash zeds that will be spawned in each wave and as
a game total assuming all waves cleared (boss not included).