Improved CDSpawn* command input handling
This commit adds better validity checking and error handling on
the user-provided parameters to the CDSpawnDetails and
CDSpawnSummaries console commands.

This commit also adds multiple formatting options to the
CDSpawnDetails command, so that the user can control the level
of abbreviation used when formatting zed type names in a cycle.