Add validity checks for custom squad elements
This commit makes several changes, all of which are concerned with
validity checking and error feedback on user-provided squad spawn
schedules that get parsed out of the config file.

* Check for all error cases at the squad element level: does
  it specify a valid number of zeds [1-6], does it specify a
  valid zed name?

* The set of acceptable zed names has been tightened.  This is
  still handled by the GetZedType method in CD_Survival.
  Basically, any two-character prefix of a zed name is accepted,
  along with certain one-character exceptions (e.g. F and G),
  and a few multi-character special cases (e.g. GF and ClotA).
  Zed name parsing is still completely case-insensitive.

  The older version of this method, adapted from the parser for
  the SpawnZed command, was more generous in what it accepted.
  For instance, "FTHOETHUOETHUTO" would be accepted and converted
  to a fleshpound.  This is no longer accepted now that CD only
  allows strict prefixes or exact zed names.

* If a squad element can't be parsed, print a related error
  message with the invalid string in question, accompanied by
  the wavenumber, squadnumber, and elementnumber to help the
  user figure out where the error lies in their config file.

Error handling for this feature is still not complete.  One big
thing that must be checked (but isn't yet) is making sure that
the number of waves defined in the config file matches the number
of waves in the currently selected gamelength, when the user has
told the mod to defer to the config file for spawn info.  It
might also be worth checking for empty squads.  Empty elements
are an error in this commit, but empty squads are not.