Refactoring special-zed remap logic
There are three places where the current CD code checks for
CDPawn_ZedCrawler, a crawler subclass that is configured to never
spawn albino traits, and when found, returns a reference to the
equivalent base-game class KFPawn_ZedCrawler.  This happens just
before the class reference touches the stats system or the
localization code.  The localization system has the crawler class
effectively hardcoded.

This commit unifies all three of these places into a single
static method in CD_Survival.  This would make it easy to add
other zed subclasses similar to the crawler one, since it's
DRY now.