Refactor CDSpawnDetails commands
It seems like an optional string argument on an exec function
causes the unrealscript to smash the stack and completely hang
the entire KF2 executable (even if the method in question
returns before executing any other instructions).  I've refactored
this into two methods, one that prints short details and one
that prints full details.  Tiny was barely differentiated from
short and is not currently executable.