add AlbinoAlphas, new spawncycle error check
This adds the AlbinoAlphas option alongside the existing AlbinoCrawlers
option.  This also modifies the spawn manager so that AlbinoAlphas and
AlbinoCrawlers are only considered when not using a SpawnCycle.

This adds a new consistency check to the SpawnCycle parser: each wave
must have at least one squad.

It also makes a few cosmetic changes to logging output.