Refactor chat commands
This commit does not change the behavior of CD's chat commands.

It does make their internal implementation a lot tidier. A bunch
of boilerplate repetition has been cut out in favor of a table-driven
approach.  I've stopped short of implementing a hash map from
command name strings to their delegate functions (which is the natural
way to do this), instead opting for a linear scan for simplicity.
CD effectively did a linear scan over CD chat command names before
this commit, too, although it was implemented as a huge/ugly if-elseif
block.

This commit also adds a method that can automatically generate a
helpful summary of chat command names and descriptions, but this
string appears to be too large for some intermediate component between
the server and client -- maybe the chat system is dropping it?  It
would be too big for the chat window anyway, I had only intended to
print it to the client's console.  This command is disabled for now.