Automate deploy-to-workshop workflow
This commit adds a hacked-up script that builds CD using

  KFEditor make -full -unattended

then, when the "deploy" argument is given, publishes it to
the workshop under either the public CD workshop item if
the git work dir is clean, or a rarely-used private CD
beta workshop item.  The beta item is not generally useful
yet because it uploads an artifact with the same name as
the true one, and that artifact defines the same gametype,
so I have to tediously clean it from my system every time
I switch in and out of beta.  It's just something I use in
cases like this, where I'm changing my publish-to-workshop
flow.

This uses the WorkshopUserTool.exe tool that comes with
the KF2 SDK to publish via scripting.