Add CDChatHelp console command
Also added a !cdhelp chat command instructing the user to
run this console command instead.  It generates dozens of lines
of output, far exceeding the 260 char broadcast message limit
imposed by BroadcastHandler's AllowsBroadcast function.  Besides,
there's no solid reason for a help request to go through the
server.  The client and server can be expected to run in lockstep
versions.  The only potential reason I see is to be able to
automatically filter the list of commands based on the effective
authorization level of the requesting user, but that's
marginal at best.  I've opted to instead just list all commands
and print the required auth level alongside each one.

Note: I tried to write a custom unrealscript commandlet to
programmatically iterate over the chat command definitions and
dump a helptext in markdown format for easy inclusion into the
docs, but the engine does not seem to scan outside the install
directory for commandlets, so mods are apparently unable to run
their own custom commandlets (without copying files into the
shared install directory, which is so hacky that I don't want to
do it).  I will probably end up using CDChatHelp's output, copied
from the logfile into a markdown file, perhaps with an intemediate
script that dresses up the space-indents with markdown bullet
points or something.