Initial commit of cohort support
This commit introduces the idea of a spawn cohort, a set of
zero or more squads that are all spawned simultaneously.

There are two options related to this concept:

* CohortSize
* MinSpawnInterval

MinSpawnInterval controls the SpawnManager wakeup that used to
be hardcoded to 1 second.

CohortSize is supposed to be the maximum number of zeds allowed
to spawn in a single cohort, but currently this is not exactly
true.  Rather, the spawner stops when it discovers that the last
squad it spawned caused the current cohort to exceed CohortSize,
resulting in cohorts slightly larger than, rather than less than
or equal to, specified size.  I will probably rework that so
that CohortSize is interpreted strictly ("less than or equal
to").

With CohortSize=48 and MinSpawnInterval=0.250, over 154 zeds
spawn within one second when starting a solo game on outpost.

Finally, I noticed that with huge numbers of zeds on the map,
teleporting is even more annoying than usual.  Not only is
teleporting happening pretty much continuously (multiple times
a second) with 100+ zeds, but the spawnvolume selection related
to teleporting seems to actually contribute to poor server
performance.  I added a boolean var controlling whether zeds
are allowed to teleport when zeds want to get "closer" to
players.  This is only half-implemented: the value is currently
hardcoded to disable teleporting and cannot be configured.

I think this still allows zeds to teleport when stuck.  I need
to check that, along with plumbing in a proper config option.