Add CohortSize & MinSpawnInterval chat commands
Also refactored the chat command setup code a bit.  This is still
bleeding all over CD_Survival and full of redundant copy-paste code.
It didn't matter when it was just a few command-line options, but
the addition of chat commands has really accelerated the code bloat
associated with each setting.

I've thought about a few potential approaches, including macros
(the webadmin code does this, and it looks horrendous), trying to
abstract it behind one or more structs, where each struct instance
represents an option and all of the logic is table driven, or doing
basically the same thing as the struct approach except using classes
and an interface.  I'm pretty sure the macro approach leads to the
dark side, but maybe one of the others is worth doing.  Still, all
of the approaches leave something to be desired, because I have to
support at least four different option types, and because the values
have to be either duplicated or read from/written to variables that
must live in CD_Survival for config compatibility purposes.

For the time being, I'm limping along with copy-paste.