Make CohortSize act as a hard limit
Before this commit, CohortSize was a retrospective soft limit;
the spawner would keep spawning squads until it exceeded CohortSize,
then it would stop.

After this commit, CohortSize is a preemptive hard limit.  The
spawner can only spawn up to CohortSize zeds in any given attempt.
If this results in a squad being fragmented, the remainder goes
into LeftoverSpawnSquad like usual.

I implemented this by overriding ShouldAddAI() to be Cohort-aware,
which allowed me to simplify some of the other code in
CDSpawnManager.