Work-in-progress commit of FakePlayers scaling
This commit converts FakePlayers from an int config var to
a string config var.  Setting it to an int works as before;
setting it to "ini" makes CD look for an additional array
of config vars named FakePlayerDefs.  Each FakePlayerDef line
is a comma-separated list of ints (secretly, the parser accepts
any float, but ints are a valid input subset).  The first
comma-separated value on a line specifies FakePlayers with one
human, second comma-separated value on a line specifies
FakePlayers with two humans, etc.  Each line controls settings
for one wave.  The first FakePlayersDefs line controls wave 1,
the second FakePlayersDefs line controls wave 2, etc.  If the
GameLength has more waves than there are lines, then the impl
just reuses the last line until the game ends.  Likewise, if
there are more humans than there are comma-separated entries
on a line, then the impl uses the last entry on the line.

The matrix specified by FakePlayers lines and their CSV
contents may be ragged.

This is only just baaarely tested, and probably missing a
bunch of edge-case checking and error messaging, but it does
appear to work in simple solo testing on hillside in a short
game.  I also intend to add another supported string value,
bilinear:<stuff>, where stuff describes a function that is
the product of wavenum and numplayers, but that's completely
unimplemneted right now.  I did leave room to plug it in
later without much trouble by hiding the relevant code behind
an interface, but there's just no impl for this syntax yet.
It is mentioned in the help string though... that's not great,
in retrospect.  Oh well, this is a WIP commit.

These changes are currently tightly coupled to FakePlayers.
I considered several ways to go about implementing this in an
abstract, table-driven fashion, where I could just declare a
datastructure describing an option and its datatype and write
general logic to parse a matrix for its values, do all the
bounds checks and error messaging, etc.  However, pretty much
every idea I considered beforehand would have run into trouble
writing back to the config var inside CD_Survival; and I want
to preserve all of those existing config vars for backwards
compat.

Macros are the one approach that could avoid the problem, but
I'm not willing to turn to the dark side on this.  This could
easily be solved with macros, but they would be similar in
complexity to the engines's webadmin macros, which are pretty
elaborate and terrible to read or debug.

So, I decided to just start on the straightforward approach
committed here, then work on generalizing it to other int vars
and finally to other float vars later.  I did cheat a bit by
making a general float parser and then casting to int though.
There will still be some duplicated, boilerplate-y code with
this approach (even after generalization) but at least it's
not macro hell.