Bring 3 float options under automatic tuning
This commit converts CD_RegulatedOption's internal state
and function signatures from using ints to using floats.

The existing 3 settings (FP, MM, CS) that used int vars atop
CD_RegulatedOption now cast between int and float.  There's
also a new virtual method on CD_RegulatedOption that it uses
to convert a float to a string when presenting it to the user,
so that the float internals can be hidden for int options.

This commit also reworks the ZTSpawnSlowdown calculation.
It now uses Lerp, similar to how KFGameInfo.TickZedTime uses
Lerp.  This implementation is much less nutty than the old
one.

A subtler change in this commit relates to the spawnmanager
wakeup timer.  This is now reset every time a wave starts.
The lack of reset is unnoticeable when using the default
MinSpawnInterval (1s) or lower, but with very high MSI (e.g.
30s), this can make a substantial difference.  Without this
reset, the timer just starts once at the beginning of the
game, before players even ready up, and it could get arbitrarily
into or out of phase with wave starts.  Now the first wave
spawn opportunity is always delayed by MinSpawnInterval or
the 5s initial TimeUntilNextSpawn value that's still
hardcoded in KFAISpawnManager.SetupNextWave, whichever is
less.