Rename RegulatedOption -> ProgrammableSetting
A RegulatedOption used to rely on a SettingRegulator.
SettingRegulator's functionality and interface is
totally unchanged, except it is now called a ValueProgram.

So now we have ProgrammableSettings that, when configured
to use a matrix of KFGame.ini literals provided by the user,
depends on an internal ValueProgram.

The verb "regulate" still lives on in some method names
and comments.