Add debug command for programmable settings
Normally, programmable settings only consider NumPlayers, which
does not appear to include spawnhumanpawn because the counter is
incremented according to steamlogins for some reason.

This command adds a console command

   CDSetDebugExtraProgramPlayers <int>

This value is normally zero.  It is added to the NumPlayers
count when programming settings.  This is useful for solo
testing programs that scale with playercount.