Handle CDEcho messages based on length
Before this commit, CDEcho chat messages (the little
cyan text sent to global chat by CD at wave start or
in response to chat commands) went through the same
logic path used by client chat messages.

This logic path includes a 260 char limit, enforced
on the server, that does not seem to reflect a hard
protocol limit, as far as I can tell.

I modified the handling of CDEcho messages to bypass
the 260 char limit.  On the client side, messages
longer than than a certain threshold are printed
only to the console, along with one line reading
"(See Console)" in the client's chat window.

This limit is actually client-side configurable
via the ChatMessageLengthThreshold ini option.
When this option is set to zero, the effective default
is 260, just like the old system.

This became necessary due to !cdinfo full occasionally
exceeding the 260 char limit, depedning on the exact
values printed.  And that's with several recently-
added options still missing from !cdinfo full.

This could probably benefit from a console command to
update the ChatMessageLengthThreshold option and save
the client's config.  It also might make sense to
limit on both char count and newlines somehow, but
I'm not sure how I would expose that to the user.