Add FakePlayersMode=add|replace
add behaves like FakePlayers always has; the configured value
is added to the count of real players to generate a sum used
for wave size computation.

replace is the new behavior.  This causes wave size computation
to only count FakePlayers, not real players.  In this mode, if
FakePlayers is set to zero (or negative somehow), then the value
1 is used instead.

Although I used wave size in the preceding paragraphs, this
setting also controls BossFP, FleshpoundFP, ScrakeFP, and
TrashFP (but instead of affecting wave size in those cases, it
of course affects zed health).  This also generalizes to
clamping minvalue=1 in replacement mode.