Move some functions out of CD_Survival
CD_Survival tends to be a kitchen-sink class when I'm
not vigilant about organization, just because it's the
most central class in the project.  This commit moves
a few accumulated kitchen-sink bits of code, like
validation logic specific to certain settings and a
string helper function, out to other classes that are
closer-related.

I'd like to eventually get CD_Survival back under 1kloc,
though I'd probably have to discount comments for that
to happen again.