Rename MinSpawnInterval to SpawnPoll
This is kind of an inferior name in terms of semantic precision,
but it makes SpawnPoll look and sound more like SpawnMod, which
may hint to new and returning CD users that SpawnPoll is on par
with SpawnMod in terms of importance, and similar in function.

I worry that "MinSpawnInterval" would sound too mysterious or
vague, resulting in players potentially shifting it down their
mental list of things to investigate and tweak.  MSI/SpawnPoll
is arguably one of the most important things they can set, and
generally has a greater impact SpawnMod, especially when
SpawnMod is near zero.