Fix SpawnCycle initialization bug
The recent commits that introduced CD_Setting and
CD_BasicSetting_SpawnCycle inadvertently broke SC initialization
when the game first loads CD.  This was a simple ordering bug:
BasicSetting_SC tried to configure the SpawnManager before the
game had a chance to call InitSpawnManager().  The Settings all
get parsed in InitGame before the manager exists.

This is fixed.

Fortunately, this never made it into a public release.