Change "bilinear:..." to "bilinear;..."
This change accommodates an apparent limitation in some part
of the engine or gameframework that can't seems to handle an
`open` console command with a parameter containing a colon.

Such a command appears to be disregarded when issued.  The
console prints a newline, but nothing else.  No map is opened
and no game is started.  I couldn't figure out how to escape
colon (if there even is a supported escape syntax).

So, I decided to retreat.  Semicolon takes the place of colon
in the bilinear specifier string.  Comma takes the old place
of semicolon (separating the function body from the optional
maximum).