Fix recap's spawnrate calc on wipes
The spawnrate calculation shown in the post-wave recap always
used WaveTotalAI as the numerator.  This variable is the total
number of spawns that could be enqueued during the entire wave;
this is the number you see on your hud as soon as the wave starts.

So, if you wiped after only part of the wave spawned, the spawnrate
would be artificially high, because it assumed the entire wave
spawned when really only part of the wave spawned.

This is fixed.