Rework SpawnCycleParser's KFP interpretation
I wasn't really satisfied with using AT_BossRandom to
signify fleshpound kings.  I did that as a sort of quick
hack to work around old assumptions in ZedNameUtils and
SpawnCycleParser that made it artifically difficult to
set a fleshpound king's EAIType to AT_FleshPound without
losing some information.

I've removed those assumptions and reworked the code so
that king fleshpounds can have any EAIType.  It's currently
using AT_FleshPound.

I suspect there is a much broader refactor lurking in here,
where I remove the if-else constructions in ZedNameUtils
and instead turn to a table-driven approach, but that's
likely to encourage a lot of disruptive (though ultimately
simplifying) tweaks to CD_SpawnManager, and I'm not ready
to puruse that right now.