Remove !cdhelp chat docs
This chat command is gone, but its docs remained.  The closest available
(though not really satisfactory) replacement is CDChatHelp in console,
but that only works in standalone and on the dedicated server.

In an abstract sense, !cdhelp should not need to talk to the server to
compute its answer and display it to the client, but there are some
technical and design limitations that make this artificially hard right
now.  In particular, a multiplayer client does not have a GameInfo, and
the central CD_ChatCommander class that catalogs chat commands and
knows how to generate help text requires a GameInfo to setup callback
functions.  This accidental complexity, not fundamental, and might get
worked around later.