Make post-wave recap ignore zedtime
All of the post-wave recap stats effectively observed zedtime,
which is crazy.  We think in terms of real time, not zed-time-adjusted
real time.  This meant that the "spawning" and "post" figures reported
could be smaller than their true external-player-with-a-stopwatch values.
Furthermore, because the spawnrate is calculated with the "spawning"
figure in the denominator, the spawnrate could be higher than its true
value.  The effect of this bug is magnified as more and more of the wave
is spent in zed time (of course, this includes commando extensions, so
zed time can potentially account for a huge chunk of the wave).

Specifically, the recap used WorldInfo.TimeSeconds, which passes
artificially slowly during zed time.

As of this commit, it uses WorldInfo.RealTimeSeconds, which does not
pass artificially slowly during zed time.

This was a really harmful bug, because the effect could sometimes be
quite subtle.

I did not see it until I started comparing external timestamps to long
waves, since the effect only becomes pronounced if it is long enough to
have lots of zed time.

I've tested this fix by:

* Playing solo hillside wave 1 on basic_moderate with 5 faked as
  commando, and going for maximum extensions
* Comparing the logfile timestamps with the post-wave recap
* Playing with a stopwatch on an auxiliary monitor and noting the
  times at which:
  (a) the wave starts
  (b) the zeds remaining counter = MaxMonsters
  (c) the wave ends
  and then comparing those stopwatch times to the recap.

Both of these tests showed agreement within one second over a three-minute
wave (three minutes of real wall-clock time, that is).