Document FakePlayers etc rename in options.md
My previous commit updated all of the existing docs to refer to
the new fakes options, but it did not leave any historical note
about the old FakePlayers etc options.  To help new users who
stumble on an old CD config, I've added a section to the docs that
mentions all of the old option names and how they map to current
option names.