Fix typo in albino_heavy
I'm guessing Kore hand-converted this from KFGame.ini SpawnCycleDefs
to a .uc source file just before sending this to me.  This is why I
prefer ini snippets!  There's no way this typo was present when he
tested, because the parse error is detected at load time, before the
game even starts.  Fortunately, this seems to be the only character
out of place.