Removing FleshpoundKing
Tripwire has rebalanced KFP to be on par with Patriarch and Hans, and altered
his implementation in so doing, breaking CD's old KFP-in-SpawnCycle support.
At this point, to reintroduce KFP into SpawnCycles, I would essentially have to
solve all of the technical and design problems related to having bosses in
normal waves.  I still think that I have neither a good design for giving the
user fine-grained control over the lumpy difficulty spike imposed by a boss
(e.g. by adding some mechanism to let the user specify reduction in other
parameters during the boss's lifetime) nor a strong case for broad enough usage
to justify the moderate to above-average time it would probably take to
implement, test, and indefinitely support.

Therefore, I'm reluctantly abandoning KFP support in SpawnCycles.  I'm not
doing this lightly, but it seems like the least flawed response to the
rebalance, at least for the intermediate term.  At least I had the foresight to
mark KFP support as experimental in the docs.